Add Apple GPU (mps) support via shared device resolver
parse_device_dtype now accepts mps (16/32-bit float dtypes only;
float64 and non-float dtypes fall back to cpu with a warning) and all
decoders and error models resolve device/dtype through it instead of
per-module copies. cuda without a GPU falls back to cpu; devices
resolve to explicit cuda:<idx>, with checkpoint validation treating
cuda and cuda:0 as equal so existing checkpoints resume.

+++ src/syndrilla/decoder/saq/saq.py
@@ -176,7 +176,7 @@ class create(nn.Module):
 
         logger.info("Creating saq decoder.")
 
-        self.device, _ = parse_device_dtype(decoding_cfg)
+        self.device, _ = parse_device_dtype(decoding_cfg, default_dtype="float32")
         self.dtype = decoding_cfg.get("dtype", "float32")
         if self.dtype not in {"float32", "float64", "bfloat16", "float16"}:
             logger.warning(
